requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0.1
django-appconf==1.0.5
django-cors-headers==3.10.1
django-imagekit==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
pilkit==2.0
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.2.0
